kernel: UserlandKernelBoundary v2
The initial version of the `UserlandKernelBoundary` trait abstracted
details about context switching on cortex-m platforms out of process.rs,
however, the trait was very much designed around the structure of how
cortex-m handles switching between privilege modes. After working on
RISC-V context switching, it is clear some of the interface does not map
to RISC-V very well.

To make `UserlandKernelBoundary` more generic, this commit makes two
general changes to the trait.

1. Rather than having a separate "getter" for retreiving which syscall
was actually called after learning that the process stopped running
because it called a syscall, we now return the syscall information
directly with the context switch reason. Arguably, this is how it should
have been done before, and this change actually has nothing to do with
making UKB more generic.

2. The Cortex-M context switch uses a special stack frame created by the
SVC call. This approach is not used in RISC-V. Therefore, the
terminology around "popping" and "pushing" stack frames doesn't make
sense in all cases. Also, having `pop_stack_frame()` at all was a bit of
blur between the kernel loop and the context switch interface details as
pop was only called with a yield().

What is generic is the idea that we want to be able to call a function in
the process that the process should execute when it starts (or resumes)
running. That function takes over the old "push" function. "pop" has
been removed. I believe that this new interface will be implementable in
RISC-V, as well as Cortex-M, and actually improves the boundary between
the kernel loop and the context switching code.

This commit also then updates process.rs and sched.rs to use this new
interface. This required some minor changes to process.rs as these
functions are essentially passed through process.rs from the kernel
sched loop to the UKB code.